requirements.txt
asgiref==3.8.1
Django==4.2
djangorestframework==3.14.0
pytz==2024.2
sqlparse==0.5.1
typing_extensions==4.12.2
